Update resumeBuilder.js
I modified some missing semicolons and applied JSbeautifier

However, in regard to below comments;

"Creating global 'for' variable. Should be 'for (var job ...'.
The body of a for in should be wrapped in an if statement to filter unwanted properties from the prototype."

I do not agree with it. I just followed Udacity Style Guide.